Simple FGO Servant Card Simulator
This is a simple code where you can make GET requests to the /draw-cards endpoint. You can simulate 3 consecutive turns. I cannot guarantee that the card drawing logic in FGO works exactly like this, and I'm not considering scenarios such as a Servant dying in the middle of the turns, as the deck is reshuffled when that happens. I literally created this code after trying 40 times in a row to get 3 Jalter cards on my second turn, something that seemed impossible.